license_server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-telegram-bot==20.8
requests==2.31.0
